separate out errors
