+++ src/connectors/BaseConnector.ts
@@ -7,7 +7,7 @@ import { options, hpcConfig, SSH, callableFunction } from "../definitions";
 import FileUtil from "../helpers/FolderUtil";  // shouldn't this be registerUtil?
 import * as Helper from "../helpers/Helper";
 import BaseMaintainer from "../maintainers/BaseMaintainer";
-import { ConnectorError } from "../utils/errors";
+import { ConnectorError } from "../definitions";
 
 import connectionPool from "./ConnectionPool";
